rename interfaces to spec-types
