+++ src/custom-operations/apply-traits.ts
@@ -3,7 +3,7 @@ import { JSONPath } from 'jsonpath-plus';
 import { xParserOriginalTraits } from '../constants';
 import { mergePatch } from '../utils';
 
-import type { v2 } from "../interfaces";
+import type { v2 } from "../spec-types";
 
 const v2TraitPaths = [
   // operations
