+++ src/models/v2/contact.ts
@@ -5,7 +5,7 @@ import { extensions } from './mixins';
 import type { ContactInterface } from "../contact";
 import type { ExtensionsInterface } from "../extensions";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class Contact extends BaseModel<v2.ContactObject> implements ContactInterface {
   hasName(): boolean {
