+++ src/models/v2/correlation-id.ts
@@ -5,7 +5,7 @@ import { hasDescription, description, extensions } from './mixins';
 import type { CorrelationIdInterface } from "../correlation-id";
 import type { ExtensionsInterface } from "../extensions";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class CorrelationId extends BaseModel<v2.CorrelationIDObject> implements CorrelationIdInterface {
   hasDescription(): boolean {
