+++ src/models/v2/extension.ts
@@ -2,7 +2,7 @@ import { BaseModel } from "../base";
 
 import type { ExtensionInterface } from "../extension";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class Extension extends BaseModel<v2.SpecificationExtension, { name: string }> implements ExtensionInterface {
   name(): string {
