+++ src/models/v2/external-docs.ts
@@ -5,7 +5,7 @@ import { hasDescription, description, extensions } from "./mixins";
 import type { ExternalDocumentationInterface } from '../external-docs';
 import type { ExtensionsInterface } from "../extensions";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class ExternalDocumentation extends BaseModel<v2.ExternalDocumentationObject> implements ExternalDocumentationInterface {
   url(): string {
