+++ src/models/v2/license.ts
@@ -5,7 +5,7 @@ import { extensions } from './mixins';
 import type { ExtensionsInterface } from "../extensions";
 import type { LicenseInterface } from "../license";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class License extends BaseModel<v2.LicenseObject> implements LicenseInterface {
   name(): string {
