+++ src/models/v2/message-example.ts
@@ -5,7 +5,7 @@ import { extensions } from './mixins';
 import type { ExtensionsInterface } from "../extensions";
 import type { MessageExampleInterface } from "../message-example";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class MessageExample extends BaseModel<v2.MessageExampleObject> implements MessageExampleInterface {
   hasName(): boolean {
