+++ src/models/v2/message.ts
@@ -18,7 +18,7 @@ import type { ServersInterface } from "../servers";
 import type { ServerInterface } from "../server";
 import type { SchemaInterface } from "../schema";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class Message extends MessageTrait<v2.MessageObject> implements MessageInterface {
   hasPayload(): boolean {
