+++ src/models/v2/server-variable.ts
@@ -5,7 +5,7 @@ import { hasDescription, description, extensions } from './mixins';
 import type { ExtensionsInterface } from '../extensions';
 import type { ServerVariableInterface } from '../server-variable';
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class ServerVariable extends BaseModel<v2.ServerVariableObject, { id: string }> implements ServerVariableInterface {  
   id(): string {
