+++ src/models/v2/tag.ts
@@ -6,7 +6,7 @@ import type { ExtensionsInterface } from "../extensions";
 import  type{ ExternalDocumentationInterface } from "../external-docs";
 import type { TagInterface } from "../tag";
 
-import type { v2 } from "../../interfaces";
+import type { v2 } from "../../spec-types";
 
 export class Tag extends BaseModel<v2.TagObject> implements TagInterface {
   name(): string {
